requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
frozenlist==1.3.0
future==0.18.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.0
opencv-python==4.6.0.66
requests==2.28.1
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
